+++ src/components/table/table-cell/table-cell.component.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import StyledTableCell from "./table-cell.style";
 import Date from "../../../__experimental__/components/date";
-import TextArea from "../../../__experimental__/components/textarea";
+import TextArea from "../../textarea";
 import { validProps } from "../../../utils/ether";
 import tagComponent from "../../../utils/helpers/tags";
 import OptionsHelper from "../../../utils/helpers/options-helper";
