+++ src/components/table/table-cell/table-cell.spec.js
@@ -11,7 +11,7 @@ import { rootTagTest } from "../../../utils/helpers/tags/tags-specs";
 import BaseTheme from "../../../style/themes/base";
 import mintTheme from "../../../style/themes/mint";
 import Date from "../../../__experimental__/components/date";
-import TextArea from "../../../__experimental__/components/textarea";
+import TextArea from "../../textarea";
 import TextBox from "../../textbox";
 import { assertStyleMatch } from "../../../__spec_helper__/test-utils";
 import tableSizes from "../table-sizes.style.js";
