+++ src/components/textarea/character-count/index.js
