+++ src/components/textarea/textarea.stories.mdx
@@ -1,12 +1,12 @@
 import { useState } from "react";
 import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 
 import Textarea from ".";
 import { OriginalTextarea } from "./textarea.component";
-import Box from "../../../components/box";
+import Box from "../box";
 
-<Meta title="Experimental/Textarea" parameters={{ info: { disable: true } }} />
+<Meta title="Design System/Textarea" parameters={{ info: { disable: true } }} />
 
 # Textarea
 
@@ -23,7 +23,7 @@ import Box from "../../../components/box";
 ## Quick Start
 
 ```javascript
-import Textarea from "carbon-react/lib/__experimental__/components/textarea";
+import Textarea from "carbon-react/lib/components/textarea";
 ```
 
 ## Examples
